Move MainChart into it's own control
